refacator: 중간 병합에 따른 프로젝트 구조 재조정
