--- src/main/java/com/goorm/team9/icontact/api/location/dto/LocationRequest.java
@@ -1,15 +1,0 @@
-package com.goorm.team9.icontact.api.location.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class LocationRequest {
-    private Long id;
-    private double latitude;
-    private double longitude;
-    private String interest;
-}
