--- src/main/java/com/goorm/team9/icontact/chat/entity/ChatMessageType.java
@@ -1,5 +1,0 @@
-package com.goorm.team9.icontact.chat.entity;
-
-public enum ChatMessageType {
-    CHAT, JOIN, LEAVE
-}
