--- src/main/java/com/goorm/team9/icontact/chat/repository/ChatJoinRepository.java
@@ -1,14 +1,0 @@
-package com.goorm.team9.icontact.chat.repository;
-
-import com.goorm.team9.icontact.chat.entity.ChatJoin;
-import com.goorm.team9.icontact.chat.entity.ChatRoom;
-import io.lettuce.core.dynamic.annotation.Param;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ChatJoinRepository extends JpaRepository<ChatJoin, Long> {
-    Optional<ChatJoin> findByChatRoomAndClientId(
-            @Param("chatRoom") ChatRoom chatRoom,
-            @Param("clientId") Long clientId);
-}
