--- src/main/java/com/goorm/team9/icontact/chat/repository/ChatRequestRepository.java
@@ -1,11 +1,0 @@
-package com.goorm.team9.icontact.chat.repository;
-
-import com.goorm.team9.icontact.chat.entity.ChatRequest;
-import com.goorm.team9.icontact.chat.entity.RequestStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ChatRequestRepository extends JpaRepository<ChatRequest, Long> {
-    Optional<ChatRequest> findByIdAndStatus(Long requestId, RequestStatus status);
-}
