+++ src/main/java/com/goorm/team9/icontact/domain/chat/dto/ChatRoomRequest.java
@@ -1,0 +1,11 @@
+package com.goorm.team9.icontact.domain.chat.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ChatRoomRequest {
+    private String senderNickname;
+    private String receiverNickname;
+}
