+++ src/main/java/com/goorm/team9/icontact/domain/chat/entity/ChatMessageType.java
@@ -1,0 +1,5 @@
+package com.goorm.team9.icontact.domain.chat.entity;
+
+public enum ChatMessageType {
+    CHAT, JOIN, LEAVE
+}
