+++ src/main/java/com/goorm/team9/icontact/domain/chat/repository/ChatMessageRepository.java
@@ -1,0 +1,11 @@
+package com.goorm.team9.icontact.domain.chat.repository;
+
+import com.goorm.team9.icontact.domain.chat.entity.ChatMessage;
+import com.goorm.team9.icontact.domain.chat.entity.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+    List<ChatMessage> findByChatRoom(ChatRoom chatRoom);
+}
