+++ src/main/java/com/goorm/team9/icontact/domain/location/dto/LocationRequest.java
@@ -1,0 +1,15 @@
+package com.goorm.team9.icontact.domain.location.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class LocationRequest {
+    private Long id;
+    private double latitude;
+    private double longitude;
+    private String interest;
+}
